common delete group functionality
