+++ frontend/src/context/AuthContext.js
@@ -1,12 +1,14 @@
 import { createContext, useState, useEffect } from "react";
 import { jwtDecode } from "jwt-decode";
 import { useHistory } from "react-router-dom";
+import Swal from "sweetalert2";
+import withReactContent from "sweetalert2-react-content";
 const swal = require("sweetalert2");
 
 const AuthContext = createContext();
 
 export default AuthContext;
-
+const MySwal = withReactContent(Swal);
 export const AuthProvider = ({ children }) => {
   const [authTokens, setAuthTokens] = useState(() =>
     localStorage.getItem("authTokens")
@@ -300,6 +302,42 @@ const verifyOldPassword = async (email, oldPassword) => {
     return false;
   }
 };
+ const confirmAndDeleteGroup = async (groupId) => {
+        const result = await MySwal.fire({
+          title: "Are you sure?",
+          text: "This group will be permanently deleted.",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#d33",
+          cancelButtonColor: "#3085d6",
+          confirmButtonText: "Yes, delete it!",
+        });
+      
+        if (result.isConfirmed) {
+          try {
+            const response = await fetch(`http://localhost:8080/api/groups/${groupId}`, {
+              method: "DELETE",
+              headers: {
+                "Content-Type": "application/json",
+                "Authorization": `Bearer ${JSON.parse(localStorage.getItem("authTokens"))}`,
+              },
+            });
+      
+            if (response.ok) {
+              Swal.fire("Deleted!", "Group has been deleted.", "success");
+              return true;
+            } else {
+              const errorText = await response.text();
+              throw new Error(errorText || "Failed to delete group.");
+            }
+          } catch (error) {
+            Swal.fire("Error", error.message, "error");
+            return false;
+          }
+        }
+        return false;
+      };
+
 
 
   // Run Check on Component Load
@@ -347,6 +385,7 @@ const verifyOldPassword = async (email, oldPassword) => {
     verifyOTP,
     resetPassword,
     verifyOldPassword,
+    confirmAndDeleteGroup,
 
 
   };
